remove unnecessary /cfi/max-ratings redirect
Given that VRO will need to rebuild the API gateway in the VA.gov cloud, we can make the path whatever we'd like. While [the gateway currently in the LHDI cloud uses `Path=/cfi/**`](https://github.com/department-of-veterans-affairs/abd-vro-api-gateway/blob/main/src/main/resources/application-local.yml#L36-L39), we are not bound to that. A redirect will not be necessary either so that I will delete that portion of the code.

Here's an issue to add this API to a VA.gov cloud gateway:
https://github.com/department-of-veterans-affairs/abd-vro/issues/3850